% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{AdministratorWebToken}
\alias{AdministratorWebToken}
\title{AdministratorWebToken Object}
\usage{
AdministratorWebToken(token = NULL)
}
\arguments{
\item{token}{An opaque token to be passed to the Play front-end to generate an iframe}
}
\value{
AdministratorWebToken object
}
\description{
AdministratorWebToken Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A token authorizing an admin to access an iframe.
}
