% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{AdministratorWebTokenSpec}
\alias{AdministratorWebTokenSpec}
\title{AdministratorWebTokenSpec Object}
\usage{
AdministratorWebTokenSpec(parent = NULL, permission = NULL)
}
\arguments{
\item{parent}{The URI of the parent frame hosting the iframe}

\item{permission}{The list of permissions the admin is granted within the iframe}
}
\value{
AdministratorWebTokenSpec object
}
\description{
AdministratorWebTokenSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specification for a token used to generate iframes. The token specifies what data the admin is allowed to modify and the URI the iframe is allowed to communiate with.
}
\seealso{
Other AdministratorWebTokenSpec functions: \code{\link{enterprises.createWebToken}}
}
