% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\docType{package}
\name{androidenterprise_googleAuthR}
\alias{androidenterprise_googleAuthR}
\alias{androidenterprise_googleAuthR-package}
\title{Google Play EMM API
Manages the deployment of apps to Android for Work users.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:34:46
filename: /Users/mark/dev/R/autoGoogleAPI/googleandroidenterprisev1.auto/R/androidenterprise_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}
}

