% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{AppRestrictionsSchema}
\alias{AppRestrictionsSchema}
\title{AppRestrictionsSchema Object}
\usage{
AppRestrictionsSchema(restrictions = NULL)
}
\arguments{
\item{restrictions}{The set of restrictions that make up this schema}
}
\value{
AppRestrictionsSchema object
}
\description{
AppRestrictionsSchema Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the list of app restrictions available to be pre-configured for the product.
}
