% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{AppRestrictionsSchemaChangeEvent}
\alias{AppRestrictionsSchemaChangeEvent}
\title{AppRestrictionsSchemaChangeEvent Object}
\usage{
AppRestrictionsSchemaChangeEvent(productId = NULL)
}
\arguments{
\item{productId}{The id of the product (e}
}
\value{
AppRestrictionsSchemaChangeEvent object
}
\description{
AppRestrictionsSchemaChangeEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event generated when a new app version is uploaded to Google Play and its app restrictions schema changed. To fetch the app restrictions schema for an app, use Products.getAppRestrictionsSchema on the EMM API.
}

