% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{AppRestrictionsSchemaRestriction}
\alias{AppRestrictionsSchemaRestriction}
\title{AppRestrictionsSchemaRestriction Object}
\usage{
AppRestrictionsSchemaRestriction(defaultValue = NULL, description = NULL,
  entry = NULL, entryValue = NULL, key = NULL, nestedRestriction = NULL,
  restrictionType = NULL, title = NULL)
}
\arguments{
\item{defaultValue}{The default value of the restriction}

\item{description}{A longer description of the restriction, giving more detail of what it affects}

\item{entry}{For choice or multiselect restrictions, the list of possible entries' human-readable names}

\item{entryValue}{For choice or multiselect restrictions, the list of possible entries' machine-readable values}

\item{key}{The unique key that the product uses to identify the restriction, e}

\item{nestedRestriction}{For bundle or bundleArray restrictions, the list of nested restrictions}

\item{restrictionType}{The type of the restriction}

\item{title}{The name of the restriction}
}
\value{
AppRestrictionsSchemaRestriction object
}
\description{
AppRestrictionsSchemaRestriction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A restriction in the App Restriction Schema represents a piece of configuration that may be pre-applied.
}
