% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{AppRestrictionsSchemaRestrictionRestrictionValue}
\alias{AppRestrictionsSchemaRestrictionRestrictionValue}
\title{AppRestrictionsSchemaRestrictionRestrictionValue Object}
\usage{
AppRestrictionsSchemaRestrictionRestrictionValue(type = NULL,
  valueBool = NULL, valueInteger = NULL, valueMultiselect = NULL,
  valueString = NULL)
}
\arguments{
\item{type}{The type of the value being provided}

\item{valueBool}{The boolean value - this will only be present if type is bool}

\item{valueInteger}{The integer value - this will only be present if type is integer}

\item{valueMultiselect}{The list of string values - this will only be present if type is multiselect}

\item{valueString}{The string value - this will be present for types string, choice and hidden}
}
\value{
AppRestrictionsSchemaRestrictionRestrictionValue object
}
\description{
AppRestrictionsSchemaRestrictionRestrictionValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A typed value for the restriction.
}

