% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ApprovalUrlInfo}
\alias{ApprovalUrlInfo}
\title{ApprovalUrlInfo Object}
\usage{
ApprovalUrlInfo(approvalUrl = NULL)
}
\arguments{
\item{approvalUrl}{A URL that displays a product's permissions and that can also be used to approve the product with the Products}
}
\value{
ApprovalUrlInfo object
}
\description{
ApprovalUrlInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information on an approval URL.
}
