% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{AppUpdateEvent}
\alias{AppUpdateEvent}
\title{AppUpdateEvent Object}
\usage{
AppUpdateEvent(productId = NULL)
}
\arguments{
\item{productId}{The id of the product (e}
}
\value{
AppUpdateEvent object
}
\description{
AppUpdateEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event generated when a new version of an app is uploaded to Google Play. Notifications are sent for new public versions only: alpha, beta, or canary versions do not generate this event. To fetch up-to-date version history for an app, use Products.Get on the EMM API.
}
