% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{AppVersion}
\alias{AppVersion}
\title{AppVersion Object}
\usage{
AppVersion(versionCode = NULL, versionString = NULL)
}
\arguments{
\item{versionCode}{Unique increasing identifier for the app version}

\item{versionString}{The string used in the Play store by the app developer to identify the version}
}
\value{
AppVersion object
}
\description{
AppVersion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This represents a single version of the app.
}
