% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{AuthenticationToken}
\alias{AuthenticationToken}
\title{AuthenticationToken Object}
\usage{
AuthenticationToken(token = NULL)
}
\arguments{
\item{token}{The authentication token to be passed to the device policy client on the device where it can be used to provision the account for which this token was generated}
}
\value{
AuthenticationToken object
}
\description{
AuthenticationToken Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An AuthenticationToken is used by the EMM's device policy client on a device to provision the given EMM-managed user on that device.
}

