% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{Collection}
\alias{Collection}
\title{Collection Object}
\usage{
Collection(collectionId = NULL, name = NULL, productId = NULL,
  visibility = NULL)
}
\arguments{
\item{collectionId}{Arbitrary unique ID, allocated by the API on creation}

\item{name}{A user-friendly name for the collection (should be unique), e}

\item{productId}{The IDs of the products in the collection, in the order in which they should be displayed}

\item{visibility}{Whether this collection is visible to all users, or only to the users that have been granted access through the 'Collectionviewers' API}
}
\value{
Collection object
}
\description{
Collection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection resource defines a named set of apps that is visible to a set of users in the Google Play Store app running on those users' managed devices. Those users can then install any of those apps if they wish (which will trigger creation of install and entitlement resources). A user cannot install an app on a managed device unless the app is listed in at least one collection that is visible to that user.Note that the API can be used to directly install an app regardless of whether it is in any collection - so an enterprise has a choice of either directly pushing apps to users, or allowing users to install apps if they want. Which is appropriate will depend on the enterprise's policies and the purpose of the apps concerned.
}
\seealso{
Other Collection functions: \code{\link{collections.insert}},
  \code{\link{collections.patch}},
  \code{\link{collections.update}}
}

