% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{collections.patch}
\alias{collections.patch}
\title{Updates a collection. This method supports patch semantics.}
\usage{
collections.patch(Collection, enterpriseId, collectionId)
}
\arguments{
\item{Collection}{The \link{Collection} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{collectionId}{The ID of the collection}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other Collection functions: \code{\link{Collection}},
  \code{\link{collections.insert}},
  \code{\link{collections.update}}
}

