% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{CollectionsListResponse}
\alias{CollectionsListResponse}
\title{CollectionsListResponse Object}
\usage{
CollectionsListResponse(collection = NULL)
}
\arguments{
\item{collection}{An ordered collection of products which can be made visible on the Google Play Store to a selected group of users}
}
\value{
CollectionsListResponse object
}
\description{
CollectionsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The collection resources for the enterprise.
}

