% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{collectionviewers.patch}
\alias{collectionviewers.patch}
\title{Adds the user to the list of those specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only such users will see the collection. This method supports patch semantics.}
\usage{
collectionviewers.patch(User, enterpriseId, collectionId, userId)
}
\arguments{
\item{User}{The \link{User} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{collectionId}{The ID of the collection}

\item{userId}{The ID of the user}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other User functions: \code{\link{User}},
  \code{\link{collectionviewers.update}},
  \code{\link{users.insert}}, \code{\link{users.patch}},
  \code{\link{users.update}}
}

