% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{CollectionViewersListResponse}
\alias{CollectionViewersListResponse}
\title{CollectionViewersListResponse Object}
\usage{
CollectionViewersListResponse(user = NULL)
}
\arguments{
\item{user}{A user of an enterprise}
}
\value{
CollectionViewersListResponse object
}
\description{
CollectionViewersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The user resources for the collection.
}

