% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{devices.setState}
\alias{devices.setState}
\title{Sets whether a device's access to Google services is enabled or disabled. The device state takes effect only if enforcing EMM policies on Android devices is enabled in the Google Admin Console. Otherwise, the device state is ignored and all devices are allowed access to Google services. This is only supported for Google-managed users.}
\usage{
devices.setState(DeviceState, enterpriseId, userId, deviceId)
}
\arguments{
\item{DeviceState}{The \link{DeviceState} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{userId}{The ID of the user}

\item{deviceId}{The ID of the device}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other DeviceState functions: \code{\link{DeviceState}}
}

