% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{DevicesListResponse}
\alias{DevicesListResponse}
\title{DevicesListResponse Object}
\usage{
DevicesListResponse(device = NULL)
}
\arguments{
\item{device}{A managed device}
}
\value{
DevicesListResponse object
}
\description{
DevicesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The device resources for the user.
}

