% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{DeviceState}
\alias{DeviceState}
\title{DeviceState Object}
\usage{
DeviceState(accountState = NULL)
}
\arguments{
\item{accountState}{The state of the Google account on the device}
}
\value{
DeviceState object
}
\description{
DeviceState Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The state of a user's device, as accessed by the getState and setState methods on device resources.
}
\seealso{
Other DeviceState functions: \code{\link{devices.setState}}
}

