% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{Enterprise}
\alias{Enterprise}
\title{Enterprise Object}
\usage{
Enterprise(administrator = NULL, id = NULL, name = NULL,
  primaryDomain = NULL)
}
\arguments{
\item{administrator}{Admins of the enterprise}

\item{id}{The unique ID for the enterprise}

\item{name}{The name of the enterprise, for example, 'Example, Inc'}

\item{primaryDomain}{The enterprise's primary domain, such as 'example}
}
\value{
Enterprise object
}
\description{
Enterprise Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Enterprises resource represents the binding between an EMM and a specific organization. That binding can be instantiated in one of two different ways using this API as follows: - For Google managed domain customers, the process involves using Enterprises.enroll and Enterprises.setAccount (in conjunction with artifacts obtained from the Admin console and the Google API Console) and submitted to the EMM through a more-or-less manual process. - For managed Google Play Accounts customers, the process involves using Enterprises.generateSignupUrl and Enterprises.completeSignup in conjunction with the managed Google Play sign-up UI (Google-provided mechanism) to create the binding without manual steps. As an EMM, you can support either or both approaches in your EMM console. See Create an Enterprise for details.
}
\seealso{
Other Enterprise functions: \code{\link{enterprises.enroll}},
  \code{\link{enterprises.insert}}
}
