% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{EnterpriseAccount}
\alias{EnterpriseAccount}
\title{EnterpriseAccount Object}
\usage{
EnterpriseAccount(accountEmail = NULL)
}
\arguments{
\item{accountEmail}{The email address of the service account}
}
\value{
EnterpriseAccount object
}
\description{
EnterpriseAccount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A service account that can be used to authenticate as the enterprise to API calls that require such authentication.
}
\seealso{
Other EnterpriseAccount functions: \code{\link{enterprises.setAccount}}
}

