% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{enterprises.getServiceAccount}
\alias{enterprises.getServiceAccount}
\title{Returns a service account and credentials. The service account can be bound to the enterprise by calling setAccount. The service account is unique to this enterprise and EMM, and will be deleted if the enterprise is unbound. The credentials contain private key data and are not stored server-side.This method can only be called after calling Enterprises.Enroll or Enterprises.CompleteSignup, and before Enterprises.SetAccount; at other times it will return an error.Subsequent calls after the first will generate a new, unique set of credentials, and invalidate the previously generated credentials.Once the service account is bound to the enterprise, it can be managed using the serviceAccountKeys resource.}
\usage{
enterprises.getServiceAccount(enterpriseId, keyType = NULL)
}
\arguments{
\item{enterpriseId}{The ID of the enterprise}

\item{keyType}{The type of credential to return with the service account}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}
}
