% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{EnterprisesSendTestPushNotificationResponse}
\alias{EnterprisesSendTestPushNotificationResponse}
\title{EnterprisesSendTestPushNotificationResponse Object}
\usage{
EnterprisesSendTestPushNotificationResponse(messageId = NULL,
  topicName = NULL)
}
\arguments{
\item{messageId}{The message ID of the test push notification that was sent}

\item{topicName}{The name of the Cloud Pub/Sub topic to which notifications for this enterprise's enrolled account will be sent}
}
\value{
EnterprisesSendTestPushNotificationResponse object
}
\description{
EnterprisesSendTestPushNotificationResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
