% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{entitlements.update}
\alias{entitlements.update}
\title{Adds or updates an entitlement to an app for a user.}
\usage{
entitlements.update(Entitlement, enterpriseId, userId, entitlementId,
  install = NULL)
}
\arguments{
\item{Entitlement}{The \link{Entitlement} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{userId}{The ID of the user}

\item{entitlementId}{The ID of the entitlement (a product ID), e}

\item{install}{Set to true to also install the product on all the user's devices where possible}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other Entitlement functions: \code{\link{Entitlement}},
  \code{\link{entitlements.patch}}
}

