% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{EntitlementsListResponse}
\alias{EntitlementsListResponse}
\title{EntitlementsListResponse Object}
\usage{
EntitlementsListResponse(entitlement = NULL)
}
\arguments{
\item{entitlement}{An entitlement of a user to a product (e}
}
\value{
EntitlementsListResponse object
}
\description{
EntitlementsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The entitlement resources for the user.
}
