% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{GroupLicense}
\alias{GroupLicense}
\title{GroupLicense Object}
\usage{
GroupLicense(acquisitionKind = NULL, approval = NULL,
  numProvisioned = NULL, numPurchased = NULL, productId = NULL)
}
\arguments{
\item{acquisitionKind}{How this group license was acquired}

\item{approval}{Whether the product to which this group license relates is currently approved by the enterprise}

\item{numProvisioned}{The total number of provisioned licenses for this product}

\item{numPurchased}{The number of purchased licenses (possibly in multiple purchases)}

\item{productId}{The ID of the product that the license is for}
}
\value{
GroupLicense object
}
\description{
GroupLicense Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group license object indicates a product that an enterprise admin has approved for use in the enterprise. The product may be free or paid. For free products, a group license object is created in these cases: if the enterprise admin approves a product in Google Play, if the product is added to a collection, or if an entitlement for the product is created for a user via the API. For paid products, a group license object is only created as part of the first bulk purchase of that product in Google Play by the enterprise admin.The API can be used to query group licenses; the available information includes the total number of licenses purchased (for paid products) and the total number of licenses that have been provisioned, that is, the total number of user entitlements in existence for the product.Group license objects are never deleted. If, for example, a free app is added to a collection and then removed, the group license will remain, allowing the enterprise admin to keep track of any remaining entitlements. An enterprise admin may indicate they are no longer interested in the group license by marking it as unapproved in Google Play.
}
