% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{GroupLicensesListResponse}
\alias{GroupLicensesListResponse}
\title{GroupLicensesListResponse Object}
\usage{
GroupLicensesListResponse(groupLicense = NULL)
}
\arguments{
\item{groupLicense}{A group license for a product approved for use in the enterprise}
}
\value{
GroupLicensesListResponse object
}
\description{
GroupLicensesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The grouplicense resources for the enterprise.
}

