% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{GroupLicenseUsersListResponse}
\alias{GroupLicenseUsersListResponse}
\title{GroupLicenseUsersListResponse Object}
\usage{
GroupLicenseUsersListResponse(user = NULL)
}
\arguments{
\item{user}{A user of an enterprise}
}
\value{
GroupLicenseUsersListResponse object
}
\description{
GroupLicenseUsersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The user resources for the group license.
}

