% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{InstallFailureEvent}
\alias{InstallFailureEvent}
\title{InstallFailureEvent Object}
\usage{
InstallFailureEvent(deviceId = NULL, failureDetails = NULL,
  failureReason = NULL, productId = NULL, userId = NULL)
}
\arguments{
\item{deviceId}{The Android ID of the device}

\item{failureDetails}{Additional details on the failure if applicable}

\item{failureReason}{The reason for the installation failure}

\item{productId}{The id of the product (e}

\item{userId}{The ID of the user}
}
\value{
InstallFailureEvent object
}
\description{
InstallFailureEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event generated when an app installation failed on a device
}

