% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{installs.delete}
\alias{installs.delete}
\title{Requests to remove an app from a device. A call to get or list will still show the app as installed on the device until it is actually removed.}
\usage{
installs.delete(enterpriseId, userId, deviceId, installId)
}
\arguments{
\item{enterpriseId}{The ID of the enterprise}

\item{userId}{The ID of the user}

\item{deviceId}{The Android ID of the device}

\item{installId}{The ID of the product represented by the install, e}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}
}
