% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{installs.patch}
\alias{installs.patch}
\title{Requests to install the latest version of an app to a device. If the app is already installed, then it is updated to the latest version if necessary. This method supports patch semantics.}
\usage{
installs.patch(Install, enterpriseId, userId, deviceId, installId)
}
\arguments{
\item{Install}{The \link{Install} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{userId}{The ID of the user}

\item{deviceId}{The Android ID of the device}

\item{installId}{The ID of the product represented by the install, e}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other Install functions: \code{\link{Install}},
  \code{\link{installs.update}}
}
