% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{InstallsListResponse}
\alias{InstallsListResponse}
\title{InstallsListResponse Object}
\usage{
InstallsListResponse(install = NULL)
}
\arguments{
\item{install}{An installation of an app for a user on a specific device}
}
\value{
InstallsListResponse object
}
\description{
InstallsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The install resources for the device.
}
