% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{LocalizedText}
\alias{LocalizedText}
\title{LocalizedText Object}
\usage{
LocalizedText(locale = NULL, text = NULL)
}
\arguments{
\item{locale}{The BCP47 tag for a locale}

\item{text}{The text localized in the associated locale}
}
\value{
LocalizedText object
}
\description{
LocalizedText Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A localized string with its locale.
}
