% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ManagedConfiguration}
\alias{ManagedConfiguration}
\title{ManagedConfiguration Object}
\usage{
ManagedConfiguration(managedProperty = NULL, productId = NULL)
}
\arguments{
\item{managedProperty}{The set of managed properties for this configuration}

\item{productId}{The ID of the product that the managed configuration is for, e}
}
\value{
ManagedConfiguration object
}
\description{
ManagedConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A managed configuration resource contains the set of managed properties that have been configured for an Android app. The app's developer would have defined configurable properties in the managed configurations schema.
}
\seealso{
Other ManagedConfiguration functions: \code{\link{managedconfigurationsfordevice.patch}},
  \code{\link{managedconfigurationsfordevice.update}},
  \code{\link{managedconfigurationsforuser.patch}},
  \code{\link{managedconfigurationsforuser.update}}
}
