% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{managedconfigurationsfordevice.delete}
\alias{managedconfigurationsfordevice.delete}
\title{Removes a per-device managed configuration for an app for the specified device.}
\usage{
managedconfigurationsfordevice.delete(enterpriseId, userId, deviceId,
  managedConfigurationForDeviceId)
}
\arguments{
\item{enterpriseId}{The ID of the enterprise}

\item{userId}{The ID of the user}

\item{deviceId}{The Android ID of the device}

\item{managedConfigurationForDeviceId}{The ID of the managed configuration (a product ID), e}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}
}

