% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{managedconfigurationsfordevice.update}
\alias{managedconfigurationsfordevice.update}
\title{Adds or updates a per-device managed configuration for an app for the specified device.}
\usage{
managedconfigurationsfordevice.update(ManagedConfiguration, enterpriseId,
  userId, deviceId, managedConfigurationForDeviceId)
}
\arguments{
\item{ManagedConfiguration}{The \link{ManagedConfiguration} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{userId}{The ID of the user}

\item{deviceId}{The Android ID of the device}

\item{managedConfigurationForDeviceId}{The ID of the managed configuration (a product ID), e}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other ManagedConfiguration functions: \code{\link{ManagedConfiguration}},
  \code{\link{managedconfigurationsfordevice.patch}},
  \code{\link{managedconfigurationsforuser.patch}},
  \code{\link{managedconfigurationsforuser.update}}
}

