% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ManagedConfigurationsForDeviceListResponse}
\alias{ManagedConfigurationsForDeviceListResponse}
\title{ManagedConfigurationsForDeviceListResponse Object}
\usage{

  ManagedConfigurationsForDeviceListResponse(managedConfigurationForDevice = NULL)
}
\arguments{
\item{managedConfigurationForDevice}{A managed configuration for an app on a specific device}
}
\value{
ManagedConfigurationsForDeviceListResponse object
}
\description{
ManagedConfigurationsForDeviceListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The managed configuration resources for the device.
}
