% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ManagedConfigurationsForUserListResponse}
\alias{ManagedConfigurationsForUserListResponse}
\title{ManagedConfigurationsForUserListResponse Object}
\usage{
ManagedConfigurationsForUserListResponse(managedConfigurationForUser = NULL)
}
\arguments{
\item{managedConfigurationForUser}{A managed configuration for an app for a specific user}
}
\value{
ManagedConfigurationsForUserListResponse object
}
\description{
ManagedConfigurationsForUserListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The managed configuration resources for the user.
}

