% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ManagedProperty}
\alias{ManagedProperty}
\title{ManagedProperty Object}
\usage{
ManagedProperty(key = NULL, valueBool = NULL, valueBundle = NULL,
  valueBundleArray = NULL, valueInteger = NULL, valueString = NULL,
  valueStringArray = NULL)
}
\arguments{
\item{key}{The unique key that identifies the property}

\item{valueBool}{The boolean value - this will only be present if type of the property is bool}

\item{valueBundle}{The bundle of managed properties - this will only be present if type of the property is bundle}

\item{valueBundleArray}{The list of bundles of properties - this will only be present if type of the property is bundle_array}

\item{valueInteger}{The integer value - this will only be present if type of the property is integer}

\item{valueString}{The string value - this will only be present if type of the property is string, choice or hidden}

\item{valueStringArray}{The list of string values - this will only be present if type of the property is multiselect}
}
\value{
ManagedProperty object
}
\description{
ManagedProperty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A managed property of a managed configuration. The property must match one of the properties in the app restrictions schema of the product. Exactly one of the value fields must be populated, and it must match the property's type in the app restrictions schema.
}
