% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ManagedPropertyBundle}
\alias{ManagedPropertyBundle}
\title{ManagedPropertyBundle Object}
\usage{
ManagedPropertyBundle(managedProperty = NULL)
}
\arguments{
\item{managedProperty}{The list of managed properties}
}
\value{
ManagedPropertyBundle object
}
\description{
ManagedPropertyBundle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A bundle of managed properties.
}
