% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{NewDeviceEvent}
\alias{NewDeviceEvent}
\title{NewDeviceEvent Object}
\usage{
NewDeviceEvent(deviceId = NULL, managementType = NULL, userId = NULL)
}
\arguments{
\item{deviceId}{The Android ID of the device}

\item{managementType}{Identifies the extent to which the device is controlled by an Android EMM in various deployment configurations}

\item{userId}{The ID of the user}
}
\value{
NewDeviceEvent object
}
\description{
NewDeviceEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event generated when a new device is ready to be managed.
}
