% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{NewPermissionsEvent}
\alias{NewPermissionsEvent}
\title{NewPermissionsEvent Object}
\usage{
NewPermissionsEvent(approvedPermissions = NULL, productId = NULL,
  requestedPermissions = NULL)
}
\arguments{
\item{approvedPermissions}{The set of permissions that the enterprise admin has already approved for this application}

\item{productId}{The id of the product (e}

\item{requestedPermissions}{The set of permissions that the app is currently requesting}
}
\value{
NewPermissionsEvent object
}
\description{
NewPermissionsEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event generated when new permissions are added to an app.
}
