% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{Notification}
\alias{Notification}
\title{Notification Object}
\usage{
Notification(appRestrictionsSchemaChangeEvent = NULL, appUpdateEvent = NULL,
  enterpriseId = NULL, installFailureEvent = NULL, newDeviceEvent = NULL,
  newPermissionsEvent = NULL, productApprovalEvent = NULL,
  productAvailabilityChangeEvent = NULL, timestampMillis = NULL)
}
\arguments{
\item{appRestrictionsSchemaChangeEvent}{Notifications about new app restrictions schema changes}

\item{appUpdateEvent}{Notifications about app updates}

\item{enterpriseId}{The ID of the enterprise for which the notification is sent}

\item{installFailureEvent}{Notifications about an app installation failure}

\item{newDeviceEvent}{Notifications about new devices}

\item{newPermissionsEvent}{Notifications about new app permissions}

\item{productApprovalEvent}{Notifications about changes to a product's approval status}

\item{productAvailabilityChangeEvent}{Notifications about product availability changes}

\item{timestampMillis}{The time when the notification was published in milliseconds since 1970-01-01T00:00:00Z}
}
\value{
Notification object
}
\description{
Notification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A notification of one event relating to an enterprise.
}
