% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{NotificationSet}
\alias{NotificationSet}
\title{NotificationSet Object}
\usage{
NotificationSet(notification = NULL, notificationSetId = NULL)
}
\arguments{
\item{notification}{The notifications received, or empty if no notifications are present}

\item{notificationSetId}{The notification set ID, required to mark the notification as received with the Enterprises}
}
\value{
NotificationSet object
}
\description{
NotificationSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A resource returned by the PullNotificationSet API, which contains a collection of notifications for enterprises associated with the service account authenticated for the request.
}

