% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{Permission}
\alias{Permission}
\title{Permission Object}
\usage{
Permission(description = NULL, name = NULL, permissionId = NULL)
}
\arguments{
\item{description}{A longer description of the permissions giving more details of what it affects}

\item{name}{The name of the permission}

\item{permissionId}{An opaque string uniquely identifying the permission}
}
\value{
Permission object
}
\description{
Permission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A permission represents some extra capability, to be granted to an Android app, which requires explicit consent. An enterprise admin must consent to these permissions on behalf of their users before an entitlement for the app can be created.The permissions collection is read-only. The information provided for each permission (localized name and description) is intended to be used in the EMM user interface when obtaining consent from the enterprise.
}

