% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{permissions.get}
\alias{permissions.get}
\title{Retrieves details of an Android app permission for display to an enterprise admin.}
\usage{
permissions.get(permissionId, language = NULL)
}
\arguments{
\item{permissionId}{The ID of the permission}

\item{language}{The BCP47 tag for the user's preferred language (e}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}
}

