% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{Product}
\alias{Product}
\title{Product Object}
\usage{
Product(appVersion = NULL, authorName = NULL, detailsUrl = NULL,
  distributionChannel = NULL, iconUrl = NULL, productId = NULL,
  productPricing = NULL, requiresContainerApp = NULL, smallIconUrl = NULL,
  title = NULL, workDetailsUrl = NULL)
}
\arguments{
\item{appVersion}{App versions currently available for this product}

\item{authorName}{The name of the author of the product (for example, the app developer)}

\item{detailsUrl}{A link to the (consumer) Google Play details page for the product}

\item{distributionChannel}{How and to whom the package is made available}

\item{iconUrl}{A link to an image that can be used as an icon for the product}

\item{productId}{A string of the form app:<package name>}

\item{productPricing}{Whether this product is free, free with in-app purchases, or paid}

\item{requiresContainerApp}{Whether this app can only be installed on devices using the Android container app}

\item{smallIconUrl}{A link to a smaller image that can be used as an icon for the product}

\item{title}{The name of the product}

\item{workDetailsUrl}{A link to the managed Google Play details page for the product, for use by an Enterprise admin}
}
\value{
Product object
}
\description{
Product Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Products resource represents an app in the Google Play store that is available to at least some users in the enterprise. (Some apps are restricted to a single enterprise, and no information about them is made available outside that enterprise.)The information provided for each product (localized name, icon, link to the full Google Play details page) is intended to allow a basic representation of the product within an EMM user interface.
}
