% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ProductApprovalEvent}
\alias{ProductApprovalEvent}
\title{ProductApprovalEvent Object}
\usage{
ProductApprovalEvent(approved = NULL, productId = NULL)
}
\arguments{
\item{approved}{Whether the product was approved or unapproved}

\item{productId}{The id of the product (e}
}
\value{
ProductApprovalEvent object
}
\description{
ProductApprovalEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event generated when a product's approval status is changed.
}

