% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ProductAvailabilityChangeEvent}
\alias{ProductAvailabilityChangeEvent}
\title{ProductAvailabilityChangeEvent Object}
\usage{
ProductAvailabilityChangeEvent(availabilityStatus = NULL, productId = NULL)
}
\arguments{
\item{availabilityStatus}{The new state of the product}

\item{productId}{The id of the product (e}
}
\value{
ProductAvailabilityChangeEvent object
}
\description{
ProductAvailabilityChangeEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event generated whenever a product's availability changes.
}
