% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ProductPermission}
\alias{ProductPermission}
\title{ProductPermission Object}
\usage{
ProductPermission(permissionId = NULL, state = NULL)
}
\arguments{
\item{permissionId}{An opaque string uniquely identifying the permission}

\item{state}{Whether the permission has been accepted or not}
}
\value{
ProductPermission object
}
\description{
ProductPermission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A product permissions resource represents the set of permissions required by a specific app and whether or not they have been accepted by an enterprise admin.The API can be used to read the set of permissions, and also to update the set to indicate that permissions have been accepted.
}
