% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ProductPermissions}
\alias{ProductPermissions}
\title{ProductPermissions Object}
\usage{
ProductPermissions(permission = NULL, productId = NULL)
}
\arguments{
\item{permission}{The permissions required by the app}

\item{productId}{The ID of the app that the permissions relate to, e}
}
\value{
ProductPermissions object
}
\description{
ProductPermissions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about the permissions required by a specific app and whether they have been accepted by the enterprise.
}
\seealso{
Other ProductPermissions functions: \code{\link{products.updatePermissions}}
}

