% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{products.approve}
\alias{products.approve}
\title{Approves the specified product and the relevant app permissions, if any. The maximum number of products that you can approve per enterprise customer is 1,000.To learn how to use Google Play for Work to design and create a store layout to display approved products to your users, see Store Layout Design.}
\usage{
products.approve(ProductsApproveRequest, enterpriseId, productId)
}
\arguments{
\item{ProductsApproveRequest}{The \link{ProductsApproveRequest} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{productId}{The ID of the product}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other ProductsApproveRequest functions: \code{\link{ProductsApproveRequest}}
}

