% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{products.list}
\alias{products.list}
\title{Finds approved products that match a query, or all approved products if there is no query.}
\usage{
products.list(enterpriseId, approved = NULL, language = NULL,
  maxResults = NULL, query = NULL, token = NULL)
}
\arguments{
\item{enterpriseId}{The ID of the enterprise}

\item{approved}{Specifies whether to search among all products (false) or among only products that have been approved (true)}

\item{language}{The BCP47 tag for the user's preferred language (e}

\item{maxResults}{Specifies the maximum number of products that can be returned per request}

\item{query}{The search query as typed in the Google Play store search box}

\item{token}{A pagination token is contained in a request''s response when there are more products}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}
}
