% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{products.updatePermissions}
\alias{products.updatePermissions}
\title{This method has been deprecated. To programmatically approve applications, you must use the iframe mechanism via the  generateApprovalUrl and  approve methods of the Products resource. For more information, see the  Play EMM API usage requirements.The updatePermissions method (deprecated) updates the set of Android app permissions for this app that have been accepted by the enterprise.}
\usage{
products.updatePermissions(ProductPermissions, enterpriseId, productId)
}
\arguments{
\item{ProductPermissions}{The \link{ProductPermissions} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{productId}{The ID of the product}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other ProductPermissions functions: \code{\link{ProductPermissions}}
}

