% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ProductsApproveRequest}
\alias{ProductsApproveRequest}
\title{ProductsApproveRequest Object}
\usage{
ProductsApproveRequest(approvalUrlInfo = NULL)
}
\arguments{
\item{approvalUrlInfo}{The approval URL that was shown to the user}
}
\value{
ProductsApproveRequest object
}
\description{
ProductsApproveRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ProductsApproveRequest functions: \code{\link{products.approve}}
}

