% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ProductSet}
\alias{ProductSet}
\title{ProductSet Object}
\usage{
ProductSet(productId = NULL, productSetBehavior = NULL)
}
\arguments{
\item{productId}{The list of product IDs making up the set of products}

\item{productSetBehavior}{The interpretation of this product set}
}
\value{
ProductSet object
}
\description{
ProductSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of products.
}
\seealso{
Other ProductSet functions: \code{\link{users.setAvailableProductSet}}
}
