% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ProductsGenerateApprovalUrlResponse}
\alias{ProductsGenerateApprovalUrlResponse}
\title{ProductsGenerateApprovalUrlResponse Object}
\usage{
ProductsGenerateApprovalUrlResponse(url = NULL)
}
\arguments{
\item{url}{A URL that can be rendered in an iframe to display the permissions (if any) of a product}
}
\value{
ProductsGenerateApprovalUrlResponse object
}
\description{
ProductsGenerateApprovalUrlResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
