% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ProductsListResponse}
\alias{ProductsListResponse}
\title{ProductsListResponse Object}
\usage{
ProductsListResponse(pageInfo = NULL, product = NULL,
  tokenPagination = NULL)
}
\arguments{
\item{pageInfo}{General pagination information}

\item{product}{Information about a product (e}

\item{tokenPagination}{Pagination information for token pagination}
}
\value{
ProductsListResponse object
}
\description{
ProductsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The matching products.
}

