% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ServiceAccount}
\alias{ServiceAccount}
\title{ServiceAccount Object}
\usage{
ServiceAccount(key = NULL, name = NULL)
}
\arguments{
\item{key}{Credentials that can be used to authenticate as this ServiceAccount}

\item{name}{The account name of the service account, in the form of an email address}
}
\value{
ServiceAccount object
}
\description{
ServiceAccount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A service account identity, including the name and credentials that can be used to authenticate as the service account.
}

