% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ServiceAccountKey}
\alias{ServiceAccountKey}
\title{ServiceAccountKey Object}
\usage{
ServiceAccountKey(data = NULL, id = NULL, type = NULL)
}
\arguments{
\item{data}{The body of the private key credentials file, in string format}

\item{id}{An opaque, unique identifier for this ServiceAccountKey}

\item{type}{The file format of the generated key data}
}
\value{
ServiceAccountKey object
}
\description{
ServiceAccountKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Credentials that can be used to authenticate as a service account.
}
\seealso{
Other ServiceAccountKey functions: \code{\link{serviceaccountkeys.insert}}
}

