% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{ServiceAccountKeysListResponse}
\alias{ServiceAccountKeysListResponse}
\title{ServiceAccountKeysListResponse Object}
\usage{
ServiceAccountKeysListResponse(serviceAccountKey = NULL)
}
\arguments{
\item{serviceAccountKey}{The service account credentials}
}
\value{
ServiceAccountKeysListResponse object
}
\description{
ServiceAccountKeysListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

