% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{SignupInfo}
\alias{SignupInfo}
\title{SignupInfo Object}
\usage{
SignupInfo(completionToken = NULL, url = NULL)
}
\arguments{
\item{completionToken}{An opaque token that will be required, along with the Enterprise Token, for obtaining the enterprise resource from CompleteSignup}

\item{url}{A URL under which the Admin can sign up for an enterprise}
}
\value{
SignupInfo object
}
\description{
SignupInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A resource returned by the GenerateSignupUrl API, which contains the Signup URL and Completion Token.
}

