% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{StoreCluster}
\alias{StoreCluster}
\title{StoreCluster Object}
\usage{
StoreCluster(id = NULL, name = NULL, orderInPage = NULL,
  productId = NULL)
}
\arguments{
\item{id}{Unique ID of this cluster}

\item{name}{Ordered list of localized strings giving the name of this page}

\item{orderInPage}{String (US-ASCII only) used to determine order of this cluster within the parent page's elements}

\item{productId}{List of products in the order they are displayed in the cluster}
}
\value{
StoreCluster object
}
\description{
StoreCluster Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Definition of a managed Google Play store cluster, a list of products displayed as part of a store page.
}
\seealso{
Other StoreCluster functions: \code{\link{storelayoutclusters.insert}},
  \code{\link{storelayoutclusters.patch}},
  \code{\link{storelayoutclusters.update}}
}
