% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{StoreLayout}
\alias{StoreLayout}
\title{StoreLayout Object}
\usage{
StoreLayout(homepageId = NULL)
}
\arguments{
\item{homepageId}{The ID of the store page to be used as the homepage}
}
\value{
StoreLayout object
}
\description{
StoreLayout Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
General setting for the Google Play for Work store layout, currently only specifying the page to display the first time the store is opened.
}
\seealso{
Other StoreLayout functions: \code{\link{enterprises.setStoreLayout}}
}

