% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{storelayoutclusters.insert}
\alias{storelayoutclusters.insert}
\title{Inserts a new cluster in a page.}
\usage{
storelayoutclusters.insert(StoreCluster, enterpriseId, pageId)
}
\arguments{
\item{StoreCluster}{The \link{StoreCluster} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{pageId}{The ID of the page}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other StoreCluster functions: \code{\link{StoreCluster}},
  \code{\link{storelayoutclusters.patch}},
  \code{\link{storelayoutclusters.update}}
}

