% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{StoreLayoutClustersListResponse}
\alias{StoreLayoutClustersListResponse}
\title{StoreLayoutClustersListResponse Object}
\usage{
StoreLayoutClustersListResponse(cluster = NULL)
}
\arguments{
\item{cluster}{A store cluster of an enterprise}
}
\value{
StoreLayoutClustersListResponse object
}
\description{
StoreLayoutClustersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The store page resources for the enterprise.
}

