% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{storelayoutpages.patch}
\alias{storelayoutpages.patch}
\title{Updates the content of a store page. This method supports patch semantics.}
\usage{
storelayoutpages.patch(StorePage, enterpriseId, pageId)
}
\arguments{
\item{StorePage}{The \link{StorePage} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{pageId}{The ID of the page}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other StorePage functions: \code{\link{StorePage}},
  \code{\link{storelayoutpages.insert}},
  \code{\link{storelayoutpages.update}}
}
