% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{StoreLayoutPagesListResponse}
\alias{StoreLayoutPagesListResponse}
\title{StoreLayoutPagesListResponse Object}
\usage{
StoreLayoutPagesListResponse(page = NULL)
}
\arguments{
\item{page}{A store page of an enterprise}
}
\value{
StoreLayoutPagesListResponse object
}
\description{
StoreLayoutPagesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The store page resources for the enterprise.
}

