% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{StorePage}
\alias{StorePage}
\title{StorePage Object}
\usage{
StorePage(id = NULL, link = NULL, name = NULL)
}
\arguments{
\item{id}{Unique ID of this page}

\item{link}{Ordered list of pages a user should be able to reach from this page}

\item{name}{Ordered list of localized strings giving the name of this page}
}
\value{
StorePage object
}
\description{
StorePage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Definition of a Google Play for Work store page, made of a localized name and links to other pages. A page also contains clusters defined as a subcollection.
}
\seealso{
Other StorePage functions: \code{\link{storelayoutpages.insert}},
  \code{\link{storelayoutpages.patch}},
  \code{\link{storelayoutpages.update}}
}

