% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{User}
\alias{User}
\title{User Object}
\usage{
User(accountIdentifier = NULL, accountType = NULL, displayName = NULL,
  id = NULL, managementType = NULL, primaryEmail = NULL)
}
\arguments{
\item{accountIdentifier}{A unique identifier you create for this user, such as 'user342' or 'asset#44418'}

\item{accountType}{The type of account that this user represents}

\item{displayName}{The name that will appear in user interfaces}

\item{id}{The unique ID for the user}

\item{managementType}{The entity that manages the user}

\item{primaryEmail}{The user's primary email address, for example, 'jsmith@example}
}
\value{
User object
}
\description{
User Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Users resource represents an account associated with an enterprise. The account may be specific to a device or to an individual user (who can then use the account across multiple devices). The account may provide access to Google Play for Work only, or to other Google services, depending on the identity model: - Google managed domain identity model requires synchronization to Google account sources (via primaryEmail). - Android for Work Accounts identity model provides a dynamic means for enterprises to create user or device accounts as needed. These accounts provide access to Google Play for Work only.
}
\seealso{
Other User functions: \code{\link{collectionviewers.patch}},
  \code{\link{collectionviewers.update}},
  \code{\link{users.insert}}, \code{\link{users.patch}},
  \code{\link{users.update}}
}

