% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{users.patch}
\alias{users.patch}
\title{Updates the details of an EMM-managed user.Can be used with EMM-managed users only (not Google managed users). Pass the new details in the Users resource in the request body. Only the displayName field can be changed. Other fields must either be unset or have the currently active value. This method supports patch semantics.}
\usage{
users.patch(User, enterpriseId, userId)
}
\arguments{
\item{User}{The \link{User} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{userId}{The ID of the user}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other User functions: \code{\link{User}},
  \code{\link{collectionviewers.patch}},
  \code{\link{collectionviewers.update}},
  \code{\link{users.insert}}, \code{\link{users.update}}
}

