% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{users.setAvailableProductSet}
\alias{users.setAvailableProductSet}
\title{Modifies the set of products a user is entitled to access.}
\usage{
users.setAvailableProductSet(ProductSet, enterpriseId, userId)
}
\arguments{
\item{ProductSet}{The \link{ProductSet} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}

\item{userId}{The ID of the user}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other ProductSet functions: \code{\link{ProductSet}}
}

