% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{UsersListResponse}
\alias{UsersListResponse}
\title{UsersListResponse Object}
\usage{
UsersListResponse(user = NULL)
}
\arguments{
\item{user}{A user of an enterprise}
}
\value{
UsersListResponse object
}
\description{
UsersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The matching user resources.
}

