% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{UserToken}
\alias{UserToken}
\title{UserToken Object}
\usage{
UserToken(token = NULL, userId = NULL)
}
\arguments{
\item{token}{The token (activation code) to be entered by the user}

\item{userId}{The unique ID for the user}
}
\value{
UserToken object
}
\description{
UserToken Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A UserToken is used by a user when setting up a managed device or profile with their managed Google Play account on a device. When the user enters their email address and token (activation code) the appropriate EMM app can be automatically downloaded.
}
