% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{Administrator}
\alias{Administrator}
\title{Google Play EMM API Objects 
Manages the deployment of apps to Android for Work users.}
\usage{
Administrator(email = NULL)
}
\arguments{
\item{email}{The administrator's email address}
}
\value{
Administrator object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:34:48
filename: /Users/mark/dev/R/autoGoogleAPI/googleandroidenterprisev1.auto/R/androidenterprise_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Administrator Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This represents an enterprise administrator who can manage the enterprise in the Google Play for Work Store.
}

