% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_objects.R
\name{Device}
\alias{Device}
\title{Device Object}
\usage{
Device(androidId = NULL, managementType = NULL)
}
\arguments{
\item{androidId}{The Google Play Services Android ID for the device encoded as a lowercase hex string}

\item{managementType}{Identifies the extent to which the device is controlled by a managed Google Play EMM in various deployment configurations}
}
\value{
Device object
}
\description{
Device Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Devices resource represents a mobile device managed by the EMM and belonging to a specific enterprise user.This collection cannot be modified via the API. It is automatically populated as devices are set up to be managed.
}
