% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidenterprise_functions.R
\name{users.insert}
\alias{users.insert}
\title{Creates a new EMM-managed user.The Users resource passed in the body of the request should include an accountIdentifier and an accountType.}
\usage{
users.insert(User, enterpriseId)
}
\arguments{
\item{User}{The \link{User} object to pass to this method}

\item{enterpriseId}{The ID of the enterprise}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidenterprise
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidenterprise)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/android/work/play/emm-api}{Google Documentation}

Other User functions: \code{\link{User}},
  \code{\link{collectionviewers.patch}},
  \code{\link{collectionviewers.update}},
  \code{\link{users.patch}}, \code{\link{users.update}}
}

